Basic message passing from JavaScript.
This is a simplistic message passing feature to allow for JavaScript
to make requests to Erlang. This requires a tight integration of the
Erlang and JavaScript code. Messages can only be sent to JavaScript
one at a time and only while JavaScript expects them. This is purely
to overcome the blocking API design of _list functions in CouchDB.